requirements.txt
Django==5.2.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
djangorestframework==3.15.0
openai==1.109.1